feat: migration of existing implementation with plugin support
This is migrating implementation from Feb 2020 with a refactoring done
by Serhii Tsymbaliuk which also added a plugin support. It is missing
few patches from the latest code. Those will be added later.

It's a copy&paste reformatted by black. There is a lot of linter issues
which will be fixed in subsequent commits.

Signed-off-by: Petr Vobornik <pvoborni@redhat.com>