fix: backport getting killed on timeout fix
When timeout happens, only the 'child' process correstponding with
currently run step should be killed.

But the whole `te` was getting killed as the kill was send to the
whole process group.

Now creating a new process group for the child process to avoid this
issue.

This was spotted when creating a test for timeout - it was killing
also pytest.

Signed-off-by: Petr Vobornik <pvoborni@redhat.com>