fix: backport of playbook extra_args fix
There was a regression which changed interpreting extra_args as dict but the
property always expected and array. The ticket which drove implementation
of the 2 commits presented an incorrect assumation that it should be used as:
  - playbook: ../sssd-qe/playbooks/add-win-users.yaml
    extra_args: '--verbose'

While the documented expected usage is actually:
    - playbook: ../sssd-qe/playbooks/add-win-users.yaml
      extra_args:
        - '--verbose'

As another possible usage is:
    - playbook: foo/bar.yaml
      extra_args:
        - '-v'
        - '--forks=50'

I.e. it is a list which can have multiple values.

But this patch extends the behavior to be able to process also a single string.
To be more robust.

Another improvement is that playbook is added after the argument list which is
more in sync with ansible-playbook help where list of playbooks is on the end.

Signed-off-by: Petr Vobornik <pvoborni@redhat.com>